roadservice/requirements.txt
Django==3.2.3
psycopg2-binary==2.8.6
django-environ==0.4.5
geopy==2.1.0
djangorestframework==3.12.4
